movie posters on activity page
-added a randomized picture to each movie activity on profile page
