+++ src/main/movie-management-frontend/src/pages/Profile/tabs/Tabs.css
@@ -28,3 +28,25 @@
 .com {
     margin-bottom: 1%;
 }
+
+.movie-card-body {
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+    width: 100%;
+}
+
+.activity-poster {
+    box-shadow: 0 10px 6px -6px #777;
+    min-width: 14vw;
+    min-height: 35vh;
+    max-height: 35vh;
+    background-size: contain;
+    background-repeat: no-repeat;
+    background-position: center;
+}
+
+.activity-info {
+    display: flex;
+    flex-direction: column;
+}
